optimize timer logic.

+++ src/codec/codec_http.cpp
@@ -71,7 +71,7 @@ namespace kim {
         goto error;                \
     }
 
-CodecHttp::CodecHttp(Log* logger, Codec::TYPE type, int keep_alive)
+CodecHttp::CodecHttp(Log* logger, Codec::TYPE type, double keep_alive)
     : Codec(logger, type), m_keep_alive(keep_alive) {
 }
 
@@ -383,7 +383,7 @@ int CodecHttp::on_header_value(http_parser* parser, const char* at, size_t len) 
         msg->set_keep_alive(atof(value.c_str()));
     } else if (header == "Connection") {
         if (value == "keep-alive") {
-            msg->set_keep_alive(-1);  // timer logic will close it.
+            msg->set_keep_alive(-1.0);  // timer logic will close it.
         } else if (value == "close") {
             msg->set_keep_alive(0.0);
         }
