+++ src/codec/codec_http.h
@@ -12,7 +12,7 @@ namespace kim {
 
 class CodecHttp : public Codec {
    public:
-    CodecHttp(Log *logger, Codec::TYPE type, int keep_alive = -1);
+    CodecHttp(Log *logger, Codec::TYPE type, double keep_alive = -1.0);
     virtual ~CodecHttp();
 
     Codec::STATUS encode(const HttpMsg &msg, SocketBuffer *sbuf);
@@ -44,7 +44,7 @@ class CodecHttp : public Codec {
     int m_encode_cnt = 0;
     int m_decode_cnt = 0;
     bool m_is_client = false;
-    int m_keep_alive = -1;
+    double m_keep_alive = -1.0;
 
     http_parser m_parser;
     http_parser_settings m_parser_setting;
